api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
openai==1.43.0
httpx==0.27.2
httpcore==1.0.5
faiss-cpu==1.8.0.post1
numpy==1.26.4
python-dotenv==1.0.1